wrap all intrinsics in #[target_feature]
Ensure that every function which uses SIMD intrinsics is marked with
either #[inline(always)] or #[target_feature].

Inlining is crucial to performance when using intrinsics, and
target-specific intrinsics can only be inlined into other functions with
the same set of features enabled. However, #[target_feature] can only be
applied to unsafe functions, and all of the functions in Multitrack's
public API are safe. We deal with this having each public safe function
be a trivial #[inline(always)] wrapper around an unsafe function marked
with #[target_feature].