implement #[dispatch] macro
The #[dispatch] macro takes a list of architectures (where all but the
last must implement Possible, and the last must implement Supported) and
a function parameterized by `A: Arch`, and generates a wrapper function
which detects at runtime the first supported architecture in the list
and then invokes the original function for that architecture.

The current implementation branches on every call; one future
optimization could be to perform feature detection only once and cache
it in a static. Another could be to store a function pointer in a static
which initially performs feature detection and then replaces itself with
a pointer to the specialization of the original function for the
detected architecture; however, this solution won't work for functions
with other generic parameters in addition to the Arch parameter (since
different generic instantiations would need to have different function
pointers).

It's possible right now to use #[cfg_attr] to specify different arch
lists for different platforms (e.g. on x86 vs. aarch64), but it would
also be nice to have a way to specify it all in a single #[dispatch]
attribute.